+++ src/traits/check_query.rs
@@ -1,6 +1,6 @@
 pub trait CheckQueryTrait {
     fn t(&self) -> String;
-    fn s(&self) -> f64;
+    fn s(&self) -> u64;
     fn r#fn(&self) -> u64;
     fn i(&self) -> u64;
     fn fp(&self) -> u64;
